chore: Enhance student record management settings with new toggle and configuration options
