--- app-modules/student-data-model/src/Filament/Pages/ManageStudentConfiguration.php
@@ -1,110 +1,0 @@
-<?php
-
-/*
-<COPYRIGHT>
-
-    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
-
-    Advising App® is licensed under the Elastic License 2.0. For more details,
-    see https://github.com/canyongbs/advisingapp/blob/main/LICENSE.
-
-    Notice:
-
-    - You may not provide the software to third parties as a hosted or managed
-      service, where the service provides users with access to any substantial set of
-      the features or functionality of the software.
-    - You may not move, change, disable, or circumvent the license key functionality
-      in the software, and you may not remove or obscure any functionality in the
-      software that is protected by the license key.
-    - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
-      to applicable law.
-    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
-      same in return. Canyon GBS® and Advising App® are registered trademarks of
-      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
-      vigorously.
-    - The software solution, including services, infrastructure, and code, is offered as a
-      Software as a Service (SaaS) by Canyon GBS Inc.
-    - Use of this software implies agreement to the license terms and conditions as stated
-      in the Elastic License 2.0.
-
-    For more information or inquiries please visit our website at
-    https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
-
-</COPYRIGHT>
-*/
-
-namespace AdvisingApp\StudentDataModel\Filament\Pages;
-
-use AdvisingApp\StudentDataModel\Models\Student;
-use AdvisingApp\StudentDataModel\Settings\ManageStudentConfigurationSettings;
-use App\Filament\Clusters\ConstituentManagement;
-use App\Models\User;
-use Filament\Actions\Action;
-use Filament\Actions\ActionGroup;
-use Filament\Forms\Components\Toggle;
-use Filament\Pages\SettingsPage;
-use Filament\Schemas\Schema;
-use UnitEnum;
-
-class ManageStudentConfiguration extends SettingsPage
-{
-    protected static string $settings = ManageStudentConfigurationSettings::class;
-
-    protected static ?string $cluster = ConstituentManagement::class;
-
-    protected static string | UnitEnum | null $navigationGroup = 'Students';
-
-    protected static ?string $navigationLabel = 'Configuration';
-
-    protected static ?int $navigationSort = 10;
-
-    public static function canAccess(): bool
-    {
-        /** @var User $user */
-        $user = auth()->user();
-
-        if (! $user->hasLicense(Student::getLicenseType())) {
-            return false;
-        }
-
-        return $user->can(['settings.view-any']);
-    }
-
-    public function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components([
-                Toggle::make('is_enabled')
-                    ->label('Enable')
-                    ->default(false),
-            ])
-            ->disabled(! auth()->user()->can('product_admin.*.update'));
-    }
-
-    public function getRedirectUrl(): ?string
-    {
-        return ManageStudentConfiguration::getUrl();
-    }
-
-    public function save(): void
-    {
-        if (! auth()->user()->can('product_admin.*.update')) {
-            return;
-        }
-
-        parent::save();
-    }
-
-    /**
-     * @return array<Action | ActionGroup>
-     */
-    public function getFormActions(): array
-    {
-        if (! auth()->user()->can('product_admin.*.update')) {
-            return [];
-        }
-
-        return parent::getFormActions();
-    }
-}
